requirements.txt
grpcio==1.45.0
grpcio-tools==1.45.0
vosk==0.3.42
PyGObject==3.42.0
rasa==3.6.4
numpy==1.22.3
tensorflow==2.12.0
tensorboard==2.12.0
keras==2.12.0